refactor(Version): Abstract into `useDesignSystemVersion` hook
